+++ src/dialogs/settingsdialog.h
@@ -2,6 +2,8 @@
 #define SETTINGSDIALOG_H
 
 #include <QDialog>
+#include <QListWidgetItem>
+#include <entities/logfilesource.h>
 
 namespace Ui {
 class SettingsDialog;
@@ -27,8 +29,32 @@ public:
 private slots:
     void on_logFileSourceTypeComboBox_currentIndexChanged(int index);
 
+    void on_logFileSourceAddButton_clicked();
+
+    void on_logFileSourceRemoveButton_clicked();
+
+    void on_logFileSourceNameLineEdit_editingFinished();
+
+    void on_logFileSourceEzPublishServerUrlLineEdit_editingFinished();
+
+    void on_logFileSourceEzPublishServerUsernameLineEdit_editingFinished();
+
+    void on_logFileSourceEzPublishServerPasswordLineEdit_editingFinished();
+
+    void on_logFileSourceLocalPathButton_clicked();
+
+    void on_logFileSourceActiveCheckBox_stateChanged(int arg1);
+
+    void on_logFileSourceListWidget_currentItemChanged(
+            QListWidgetItem *current, QListWidgetItem *previous);
+
 private:
     Ui::SettingsDialog *ui;
+    LogFileSource _selectedLogFileSource;
+
+    void setupLogFileSourceTab();
+
+    void updateLogFileSourceTab();
 };
 
 #endif // SETTINGSDIALOG_H